usb: usbpsu: Fix MISRA C-2012 Rule 10.3 violation
Rule: MISRA C-2012 Rule 10.3
Implicit conversion of "0L" from essential type
"signed 64-bit int" to different or narrower
essential type "signed 32-bit int"

Variable type casted with essential type.


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/usbpsu/src/xusbpsu.c
@@ -185,7 +185,7 @@ s32 XUsbPsu_CfgInitialize(struct XUsbPsu *InstancePtr,
 
 	(void)XUsbPsu_SetDeviceAddress(InstancePtr, 0U);
 
-	return XST_SUCCESS;
+	return (s32)XST_SUCCESS;
 }
 
 /****************************************************************************/
@@ -272,7 +272,7 @@ u8 XUsbPsu_GetLinkState(struct XUsbPsu *InstancePtr)
 
 	RegVal = XUsbPsu_ReadReg(InstancePtr, XUSBPSU_DSTS);
 
-	return XUSBPSU_DSTS_USBLNKST(RegVal);
+	return (u8)XUSBPSU_DSTS_USBLNKST(RegVal);
 }
 
 /****************************************************************************/
